add TableHeaderColumn.tsx
